# app/static/swagger.yaml
swagger: "2.0"
info:
  title: Prodaja stanova
  description: Aplikacija automatizuje poslove vezane za rad agencije za prodaju nekretnina
  version: '1.0'
  contact:
    name: Marko Mutavdzic
    email: marko.mutavdzic@factoryww.com
...
paths:
...
  /apartment/add:
    post:
      tags:
       - Apartment
      summary: Dodavanje novih stanova
      operationId: addApartment
      description: Dodavanje novih korisnika, morate biti ulogovani kao ADMIN
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: data
          description: Podatci o novom stanu
          required: true
          schema:
            type: object
            properties:
              lamella:
                type: string
                example: 4c
              address:
                type: string
                example: Belo Vrelo 54
              quadrature:
                type: number
                example: 75
              floor:
                type: integer
                example: 4
              num_rooms:
                type: number
                example: 4
              orientation:
                type: string
                example: SEVER
                enum:
                    - "ISTOK"
                    - "ZAPAD"
                    - "SEVER"
                    - "JUG"
              num_terrace:
                type: integer
                example: 1
              price:
                type: number
                example: 100000
              lowest_price:
                type: number
                example: 95000
              status:
                type: string
                example: SLOBODAN
                enum:
                    - "SLOBODAN"
                    - "REZERVISAN"
                    - "PRODAT"
              new_construction:
                type: boolean
                example: false
              in_construction:
                type: boolean
                example: false
              available_from:
                type: string
                example: '2021-12-05'
      responses:
        200:
          description:  "{'message': 'New apartment created.'}"
        400:
          description: "{'message': 'User must be ADMIN'}"
      security:
        - JWT: []
...
  /apartment/edit:
    post:
      tags:
       - Apartment
      summary: Izmena podataka o stanu
      operationId: editApartment
      description: Izmena podataka o stanu, morate biti ulogovani kao ADMIN
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: data
          description: Unesite samo podatke koje zelite da izmenite
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 1
              lamella:
                type: string
                example: 4c
              address:
                type: string
                example: Belo Vrelo 54
              quadrature:
                type: number
                example: 75
              floor:
                type: integer
                example: 4
              num_rooms:
                type: number
                example: 4
              orientation:
                type: string
                example: SEVER
                enum:
                    - "ISTOK"
                    - "ZAPAD"
                    - "SEVER"
                    - "JUG"
              num_terrace:
                type: integer
                example: 1
              price:
                type: number
                example: 100000
              lowest_price:
                type: number
                example: 95000
              status:
                type: string
                example: SLOBODAN
                enum:
                    - "SLOBODAN"
                    - "REZERVISAN"
                    - "PRODAT"
              new_construction:
                type: boolean
                example: false
              in_construction:
                type: boolean
                example: false
              available_from:
                type: string
                example: '2021-12-05'
      responses:
        200:
          description:  "{'message': 'Apartment edited.'}"
        400:
          description: "{'message': 'User must be ADMIN'}\n
                        {'message': 'Apartment with that id doesnt exists.'}"
      security:
        - JWT: []
...
  /apartments/delete:
    post:
      tags:
       - Apartment
      summary: Brisanje stanova
      operationId: deleteApartment
      description: Brisanje stanova po ID-u, morate biti ulogovani kao ADMIN
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: data
          description: ID stana
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 1
      responses:
        200:
          description:  "{'message': 'Apartment deleted'}"
        400:
          description: "{'message': 'User must be ADMIN'}\n
                        {'message': 'Apartment with that id doesnt exists.'}"
      security:
        - JWT: []
...
  /customer/add:
    post:
      tags:
       - Customer
      summary: Dodavanje novih kupaca
      operationId: addCustomer
      description: Dodavanje novih kupaca
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: data
          description: Podatci o novom kupcu
          required: true
          schema:
            type: object
            properties:
              legal_entity:
                type: string
                example: FIZICKO
                enum:
                    - "FIZICKO"
                    - "PRAVNO"
              name:
                type: string
                example: Marko Mutavdzic
              email:
                type: string
                example: marko.mutavdzic@factoryww.com
              telephone_number:
                type: string
                example: '0641470007'
              pib_jmbg:
                type: string
                example: 1806983710075
              place:
                type: string
                example: Beograd
              street:
                type: string
                example: Belo Vrelo
              num:
                type: string
                example: 54

      responses:
        200:
          description:  "{'message': 'New customer created.'}"
        400:
          description: "{'message': 'Customer with that pib_jmbg already exists'}"
  /customer/edit:
    post:
      tags:
       - Customer
      summary: Izmena podataka o kupcima
      operationId: editCustomer
      description: Izmena podataka o kupcima
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: data
          description: Unesite samo podatke koje zelite da izmenite
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 1
              legal_entity:
                type: string
                example: FIZICKO
                enum:
                    - "FIZICKO"
                    - "PRAVNO"
              name:
                type: string
                example: Marko Mutavdzic
              email:
                type: string
                example: marko.mutavdzic@factoryww.com
              telephone_number:
                type: string
                example: '0641470007'
              pib_jmbg:
                type: string
                example: 1806983710075
              place:
                type: string
                example: Beograd
              street:
                type: string
                example: Belo Vrelo
              num:
                type: string
                example: 54

      responses:
        200:
          description:  "{'message': 'Customer edited.'}"
        400:
          description: "{'message': 'Customer with that id doesnt exists.'}\n
                        {'message': 'Customer with that pib_jmbg already exists'}"
...
  /customer/delete:
    post:
      tags:
       - Customer
      summary: Brisanje kupaca
      operationId: deleteCustomer
      description: Brisanje kupca po ID-u
      produces:
        - application/json
      consumes:
        - application/jsong
      parameters:
        - in: body
          name: data
          description: Unesite ID kupca kojeg zelite da obrisete
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 1
      responses:
        200:
          description:  "{'message': 'Customer deleted.'}"
        400:
          description: "{'message': 'Customer with that id doesnt exists.'}"
  /apartment_customer/add:
    post:
      tags:
       - Apartment Customer
      summary: Dodavanje novih ponuda
      operationId: addApartmentCustomer
      description: Dodavanje novih ponuda
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: data
          description: Podatci o novoj ponudi
          required: true
          schema:
            type: object
            properties:
              apartment_id:
                type: integer
                example: 1
              customer_id:
                type: integer
                example: 1
              customer_status:
                type: string
                example: POTECNCIJALNI
                enum:
                    - "POTECNCIJALNI"
                    - "REZERVISAO"
                    - "KUPIO"
              customer_price:
                type: number
                example: 90000
              note:
                type: string
                example: nesto
              payment_method:
                type: string
                example: KES
                enum:
                    - "KES"
                    - "KREDIT"
                    - "MESOVITO"
              deposit_amount:
                type: number
                example: 10000
              contract_deadline:
                type: string
                example: '2021-12-20'
              bank:
                type: string
                example: Intesa
              loan_amount:
                type: number
                example: 800000
              cash_amount:
                type: number
                example: 10000
              contract_number:
                type: string
                example: 2021/123
              contract_date:
                type: string
                example: '2021-12-15'
      responses:
        200:
          description:  "{'message': 'New offer created'}\n
                         {'Contract message': 'Contract with name Marko-Belo Vrelo-75kvm.docx created'}"
        400:
          description: "{'message': 'Offer from that costumer for that apartment already exists'}\n
                        {'message': 'Apartment with that id doesnt exist'}\n
                        {'message': 'Apartment is already sold'}\n
                        {'message': 'Customer with that id doesnt exist'}\n
                        {'Contract message': 'Price must be approved, status is set to POTENCIJALNI'}"
...
  /apartment_customer/edit_for_sale:
    post:
      tags:
       - Apartment Customer
      summary: Izmena podataka o ponudi od strane prodavaca
      operationId: editApartmentCustomerForSale
      description: Izmena podataka o ponudi od strane prodavaca, morate biti ulogovani kao PRODAVAC
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: data
          description: Unesite samo one podatke koje zelite da izmenite
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 1
              customer_status:
                type: string
                example: POTECNCIJALNI
                enum:
                    - "POTECNCIJALNI"
                    - "REZERVISAO"
                    - "KUPIO"
              customer_price:
                type: number
                example: 90000
              note:
                type: string
                example: nesto
      responses:
        200:
          description:  "{'message':'Offer edited'}\n
                         {'Contract message': 'Contract with name Marko-Belo Vrelo-75kvm.docx created'}"
        400:
          description: "{'message': 'User must be PRODAVAC '}\n
                        {'message': 'Offer with that id doesnt exists'}\n
                        {'message': 'Apartment with that id doesnt exist'}\n
                        {'message': 'Apartment is already sold'}\n
                        {'message': 'Customer with that id doesnt exist'}\n
                        {'Contract message': 'Price must be approved, status is set to POTENCIJALNI'}"
      security:
        - JWT: []
...
  /apartment_customer/delete:
    post:
      tags:
       - Apartment Customer
      summary: Brisanje ponude
      operationId: deleteApartmentCustomer
      description: Brisanje ponuda, morate biti prijavljeni kao FINANSIJE
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: data
          description: Unesite ID ponude
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 1
      responses:
        200:
          description: "{'message': 'Offer deleted'}"
        400:
          description: "{'message': 'User must be FINANSIJE'}\n
                        {'message': 'Offer with that id doesnt exist'}"
      security:
        - JWT: []
...
securityDefinitions:
  JWT:
     type: apiKey
     in: header
     name: access-key